+++ Totoro/__init__.py
@@ -33,15 +33,3 @@ except:
 
 from sdss.utilities.Site import Site
 site = Site()
-
-from SDSSconnect import DatabaseConnection as TotoroDBConnection
-
-# Now we create the default DB connection
-databaseConnectionString = (
-    'postgresql+psycopg2://{user}:{password}@{host}:{port}/{database}'
-    .format(**config['dbConnection']))
-TotoroDBConnection(databaseConnectionString=databaseConnectionString,
-                   new=True, name='dbConnection')
-
-from Totoro.utils import checkOpenSession
-checkOpenSession()