+++ Totoro/tests/__init__.py
@@ -1,3 +1,2 @@
-
-from Totoro import TotoroDBConnection
-TotoroDBConnection('test', default=True, new=True)
+from Totoro.db import setDefaulProfile
+setDefaulProfile('test')