+++ Totoro/tests/testSetArrangement.py
@@ -14,13 +14,13 @@ Revision history:
 
 from __future__ import division
 from __future__ import print_function
-from Totoro import TotoroDBConnection
+from Totoro.db import getConnection
 from Totoro.dbclasses import fromPlateID
 from Totoro.dbclasses.plate_utils import removeOrphanedSets
 import unittest
 
 
-db = TotoroDBConnection()
+db = getConnection('test')
 session = db.Session()